requirements.txt
fastapi
uvicorn
ultralytics
gdown
pillow
python-multipart
